exynos5420: resolve compiling warnings/errors

+++ mobicore/rootpa/Code/Common/pacmp3.c
@@ -657,7 +657,7 @@ rootpaerror_t handleResponse(cmpCommandId_t commandId, CmpMessage* outResponseP,
     if(mcRet != MC_DRV_OK)
     {
         LOGE("pacmp3 handleResponse for %d registry failed %d", commandId, mcRet);
-        if(-1==mcRet)
+        if(-1==(int)mcRet)
         {
             ret = ROOTPA_ERROR_INTERNAL;
         }
